refactor: update unit tests to use new fake version location
