+++ packages/skott/test/unit/plugins/custom-resolver.spec.ts
@@ -1,7 +1,7 @@
 import * as Option from "@effect/data/Option";
 import { describe, expect, test } from "vitest";
 
-import { InMemoryFileReader } from "../../../src/filesystem/file-reader.js";
+import { InMemoryFileReader } from "../../../src/filesystem/fake/file-reader.js";
 import { InMemoryFileWriter } from "../../../src/filesystem/file-writer.js";
 import { FakeLogger } from "../../../src/logger.js";
 import {
