+++ packages/skott/test/unit/shared.ts
@@ -1,6 +1,6 @@
 import * as memfs from "memfs";
 
-import { InMemoryFileReader } from "../../src/filesystem/file-reader.js";
+import { InMemoryFileReader } from "../../src/filesystem/fake/file-reader.js";
 import { InMemoryFileWriter } from "../../src/filesystem/file-writer.js";
 import { FakeLogger } from "../../src/logger.js";
 import { kExpectedModuleExtensions } from "../../src/modules/resolvers/base-resolver.js";
