+++ packages/skott/test/unit/traversal.spec.ts
@@ -2,7 +2,7 @@ import { describe, expect, test } from "vitest";
 
 import { mountFakeFileSystem } from "./shared.js";
 
-import { InMemoryFileReader } from "../../src/filesystem/file-reader.js";
+import { InMemoryFileReader } from "../../src/filesystem/fake/file-reader.js";
 import { InMemoryFileWriter } from "../../src/filesystem/file-writer.js";
 import { FakeLogger } from "../../src/logger.js";
 import { ModuleWalkerSelector } from "../../src/modules/walkers/common.js";
