+++ packages/skott/test/unit/workspace.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "vitest";
 
-import { InMemoryFileReader } from "../../src/filesystem/file-reader.js";
+import { InMemoryFileReader } from "../../src/filesystem/fake/file-reader.js";
 import { InMemoryFileWriter } from "../../src/filesystem/file-writer.js";
 import { FakeLogger } from "../../src/logger.js";
 import { ModuleWalkerSelector } from "../../src/modules/walkers/common.js";
